+++ build.gradle
@@ -69,7 +69,7 @@ repositories {
 }
 
 application {
-    mainClassName = "duke.Launcher"
+    mainClassName = "controller.Launcher"
 }
 
 run {
