+++ src/main/java/controller/Main.java
@@ -1,5 +1,6 @@
-package duke;
+package controller;
 
+import duke.Duke;
 import java.io.IOException;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
