+++ src/main/java/controller/MainWindow.java
@@ -1,5 +1,6 @@
-package duke;
+package controller;
 
+import duke.Duke;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
