+++ storage/file.txt
@@ -1,5 +1,4 @@
 [T]|N|4
-[E]|N|as|2020-12-09|16:00
 [T]|Y|23
 [T]|N|4
 [T]|N|4
@@ -7,8 +6,6 @@
 [T]|N|helloworld
 [T]|N|helloworld2
 [T]|Y|432
-[T]|N|4
-[T]|N|43
 [T]|N|4
 [T]|N|4
 [T]|N|4
